basic/requirements.txt
middleware-apm>=2.0.0
flask==2.3.2
